export timeline or summary will export db
# message auto-generated for no-merge-commit merge:
merge linux into master

默认导出db，包括msprof.db和analysis.db

Created-by: msprofDeveloper

Match-id-087ec116abc6cfface05574a1a5a95cc19ef0a43

+++ test/msprof_cpp/analysis_ut/domain/data_process/test/soc_bandwidth_processor_utest.cpp
@@ -173,7 +173,7 @@ TEST_F(SocBandwidthProcessorUTest, TestRunShouldReturnFalseWhenProcessSingleDevi
         auto processor = SocBandwidthProcessor(path);
         auto dataInventory = DataInventory();
 
-        MOCKER_CPP(&Utils::GetDeviceIdByDevicePath).stubs().will(returnValue(UINT16_MAX));
+        MOCKER_CPP(&Utils::GetDeviceIdByDevicePath).stubs().will(returnValue(static_cast<uint16_t>(UINT16_MAX)));
         EXPECT_FALSE(processor.Run(dataInventory, PROCESSOR_NAME_SOC));
         MOCKER_CPP(&Utils::GetDeviceIdByDevicePath).reset();
     }
